+++ lib/update/index.js
@@ -1,4 +1,3 @@
-var sitepath = require ('sitepath');
 var uuid = require('node-uuid');
 var async = require('async');
 var util = require('util'),
@@ -9,7 +8,15 @@ var squel = require("squel").useFlavour('postgres'),
 var logging = require('../logging');
 var LinkedDataBox = require('linked-data-box').LinkedDataBox;
 
-var boundLogger = logging.getRootLogger('query');
+var logentry = require('./logentry'),
+    writeLogentry = logentry.writeLogentry,
+    updateLogentry = logentry.updateLogentry,
+    execLogentry = logentry.execLogentry;
+var tags = require('./tags'),
+    clearTags = tags.clearTags,
+    writeTags = tags.writeTags;
+
+var boundLogger = logging.getRootLogger('update');
 
 squel.registerValueHandler(Date, function(date) {
   // The pg driver lets us send raw dates
@@ -22,12 +29,6 @@ squel.registerValueHandler(Date, function(date) {
 * @title Update operations
 * @module update
 */
-
-function InvalidLogClass() {
-  this.message = "evt_class in log entry doesn't match anything we expect";
-}
-util.inherits(InvalidLogClass, Error);
-errs.register('update.bad_evtClass', InvalidLogClass);
 
 function ErrorWhileRollingBack() {
   this.message = "encountered an error while rolling back";
@@ -53,240 +54,6 @@ function postCommitErr(db, ctx, client, err, done, next) {
   });
 }
 
-function writeLogentry(ctx, client, done, logentry, callback) {
-  boundLogger.info('writeLogentry', {
-    ctx: ctx
-  });
-  var s = squel.insert().into('wh_log');
-  sql.setFields(s, {
-    "path": logentry.path,
-    "\"entityId\"": logentry.entityId,
-    "note": logentry.note,
-    "\"baseRevisionId\"": logentry.baseRevisionId,
-    "\"replaceRevisionId\"": logentry.replaceRevisionId,
-    "\"revisionId\"": logentry.revisionId,
-    "\"revisionNum\"": logentry.revisionNum,
-    "\"evtStart\"": logentry.evtStart,
-    "\"evtEnd\"": logentry.evtEnd,
-    "\"evtTouched\"": logentry.evtTouched,
-    "\"evtClass\"": logentry.evtClass,
-    "\"evtFinal\"": logentry.evtFinal,
-    "\"actorPath\"": logentry.actorPath,
-    "data": JSON.stringify(logentry.data)
-  });
-  var q = s.toParam();
-  q.name = 'insert_log_query';
-  client.query(q, function(err, result) {
-    callback(err, client, done, logentry);
-  });
-}
-
-function updateLogentry(ctx, client, done, logentry, callback) {
-  boundLogger.info('updateLogentry', {
-    ctx: ctx
-  });
-  var s = squel.update().table('wh_log');
-  sql.setFields(s, {
-    "path": logentry.path,
-    "\"entityId\"": logentry.entityId,
-    "note": logentry.note,
-    "\"baseRevisionId\"": logentry.baseRevisionId,
-    "\"replaceRevisionId\"": logentry.replaceRevisionId,
-    "\"revisionNum\"": logentry.revisionNum,
-    "\"evtStart\"": logentry.evtStart,
-    "\"evtEnd\"": logentry.evtEnd,
-    "\"evtTouched\"": logentry.evtTouched,
-    "\"evtClass\"": logentry.evtClass,
-    "\"evtFinal\"": logentry.evtFinal,
-    "\"actorPath\"": logentry.actorPath,
-    "data": JSON.stringify(logentry.data)
-  });
-  s.where("\"revisionId\" = ?", logentry.revisionId);
-  var q = s.toParam();
-  q.name = 'update_log_query';
-  client.query(q, function(err, result) {
-    callback(err, client, done, logentry);
-  });
-}
-
-function execLogentry(ctx, exec, client, done, logentry, callback) {
-  boundLogger.info('execLogentry begin', {
-    ctx: ctx,
-    evtClass: logentry.evtClass,
-    exec: exec
-  });
-  var s, q;
-  if (!exec) {
-    return callback(null, client, done, logentry);
-  }
-  var now = new Date();
-  if (logentry.evtClass === 'Create') {
-    var ds = squel.remove().from('wh_entity');
-    ds.where("path = ?", logentry.path);
-    ds.where("stub = true");
-    var dq = ds.toParam();
-    dq.name = 'insert_entity_query_delete';
-
-    s = squel.insert().into('wh_entity');
-    sql.setFields(s, {
-      "path": logentry.path,
-      "stub": false,
-      "\"entityId\"": logentry.entityId,
-      "\"revisionId\"": logentry.revisionId,
-      "\"revisionNum\"": logentry.revisionNum,
-      "proto": logentry.data.toData.proto,
-      "modified": logentry.data.toData.modified,
-      "created": logentry.data.toData.created,
-      "summary": JSON.stringify(logentry.data.toData.summary),
-      "data": JSON.stringify(logentry.data.toData.data),
-      "tags": JSON.stringify(logentry.data.toData.tags)
-    });
-    q = s.toParam();
-    q.name = 'insert_entity_query';
-
-    client.query(dq, function(err, result) {
-      if (err) {
-        return callback(err, client, done, logentry);
-      }
-      client.query(q, function(err, result) {
-        if (!err) {
-          logentry.evtFinal = true;
-          logentry.evtEnd = now;
-        }
-        callback(err, client, done, logentry);
-      });
-    });
-  } else if (logentry.evtClass === 'Update') {
-    s = squel.update().table('wh_entity');
-    sql.setFields(s, {
-      "stub": false,
-      "\"entityId\"": logentry.entityId,
-      "\"revisionId\"": logentry.revisionId,
-      "\"revisionNum\"": logentry.revisionNum,
-      "proto": logentry.data.toData.proto,
-      "modified": logentry.data.toData.modified,
-      "created": logentry.data.toData.created,
-      "summary": JSON.stringify(logentry.data.toData.summary),
-      "data": JSON.stringify(logentry.data.toData.data),
-      "tags": JSON.stringify(logentry.data.toData.tags)
-    });
-    s.where("path = ?", logentry.path);
-    q = s.toParam();
-    q.name = 'update_entity_query';
-
-    client.query(q, function(err, result) {
-      if (!err) {
-        logentry.evtFinal = true;
-        logentry.evtEnd = now;
-      }
-      callback(err, client, done, logentry);
-    });
-  } else if (logentry.evtClass === 'Delete') {
-    s = squel.update().table('wh_entity');
-    sql.setFields(s, {
-      "stub": true,
-      "\"entityId\"": logentry.entityId,
-      "\"revisionId\"": logentry.revisionId,
-      "\"revisionNum\"": logentry.revisionNum,
-      "proto": null,
-      "modified": logentry.evtTouched,
-      "created": logentry.data.fromData.created,
-      "summary": JSON.stringify({'deleted': true}),
-      "data": JSON.stringify({}),
-      "tags": JSON.stringify({})
-    });
-    s.where("path = ?", logentry.path);
-    q = s.toParam();
-    q.name = 'update_entity_query';
-    client.query(q, function(err, result) {
-      if (!err) {
-        logentry.evtFinal = true;
-        logentry.evtEnd = now;
-      }
-      callback(err, client, done, logentry);
-    });
-  } else if (logentry.evtClass === 'Move') {
-    s = squel.update().table('wh_entity');
-    sql.setFields(s, {
-      "path": logentry.data.toPath,
-      "\"revisionId\"": logentry.revisionId,
-      "\"revisionNum\"": logentry.revisionNum,
-    });
-    s.where("path = ?", logentry.path);
-    q = s.toParam();
-    q.name = 'move_entity_query';
-    client.query(q, function(err, result) {
-      if (!err) {
-        logentry.evtFinal = true;
-        logentry.evtEnd = now;
-      }
-      callback(err, client, done, logentry);
-    });
-  } else if (logentry.evtClass === 'rm3:assign') {
-    s = squel.insert().into('wh_subject_to_roles');
-    sql.setFields(s, {
-      "subject": logentry.path,
-      "role": logentry.data.role
-    });
-    q = s.toParam();
-    q.name = 'assign_subject_to_role';
-    client.query(q, function(err, result) {
-      if (!err) {
-        logentry.evtFinal = true;
-        logentry.evtEnd = now;
-      }
-      callback(err, client, done, logentry);
-    });
-  } else if (logentry.evtClass === 'rm3:deassign') {
-    s = squel.remove().from('wh_subject_to_roles');
-    s.where("subject = ?", logentry.path);
-    s.where("role = ?", logentry.data.role);
-    q = s.toParam();
-    q.name = 'remove_user_from_role';
-    client.query(q, function(err, result) {
-      if (!err) {
-        logentry.evtFinal = true;
-        logentry.evtEnd = now;
-      }
-      callback(err, client, done, logentry);
-    });
-  } else if (logentry.evtClass === 'rm3:permit') {
-    s = squel.insert().into('wh_permission_to_role');
-    sql.setFields(s, {
-      "path": logentry.data.path,
-      "role": logentry.data.role,
-      "permission": logentry.data.permission,
-      "query": logentry.data.path
-    });
-    q = s.toParam();
-    q.name = 'add_permission_to_role';
-    client.query(q, function(err, result) {
-      if (!err) {
-        logentry.evtFinal = true;
-        logentry.evtEnd = now;
-      }
-      callback(err, client, done, logentry);
-    });
-  } else if (logentry.evtClass === 'rm3:deny') {
-    s = squel.remove().from('wh_permission_to_role');
-    s.where("path = ?", logentry.data.path);
-    s.where("role = ?", logentry.data.role);
-    s.where("permission = ?", logentry.data.permission);
-    q = s.toParam();
-    q.name = 'remove_permission_from_role';
-    client.query(q, function(err, result) {
-      if (!err) {
-        logentry.evtFinal = true;
-        logentry.evtEnd = now;
-      }
-      callback(err, client, done, logentry);
-    });
-  } else {
-    var err = errs.create('update.bad_evtClass', {evtClass: logentry.evtClass});
-    callback(err, client, done, logentry);
-  }
-}
-
 exports._private = {};
 exports._private.execLogentry = execLogentry;
 
@@ -305,73 +72,6 @@ function doCleanup(db, ctx, next, err, client, done, logentry) {
   }
   done(err);
   return next(err, logentry.entityId, logentry.revisionId, logentry.revisionNum);
-}
-
-function clearTags(ctx, exec, client, done, logentry, callback) {
-  boundLogger.info('clearTags', {
-    ctx: ctx,
-    path: logentry.path,
-    exec: exec
-  });
-  if (!exec) {
-    return callback(null, client, done, logentry);
-  }
-
-  var path = logentry.path;
-  var s = squel.remove().from('wh_tag');
-  s.where("\"subjPath\" = ?", path);
-  var q = s.toParam();
-  q.name = 'clear_tags';
-  client.query(q, function(err, result) {
-    callback(err, client, done, logentry);
-  });
-}
-
-function generateOneTagInsert(subjPath, predKey, objKey, objClass, obj) {
-  var s = squel.insert().into('wh_tag');
-  sql.setFields(s, {
-      "\"subjPath\"": subjPath,
-      "\"objClass\"": objClass,
-      "\"predPath\"": predKey,
-      "\"objStr\"": objKey,
-    });
-  var q = s.toParam();
-  q.name = 'insert_tag_query';
-  return q;
-}
-
-function writeTags(ctx, exec, client, done, logentry, callback) {
-  boundLogger.info('writeTags', {
-    ctx: ctx,
-    path: logentry.path,
-    exec: exec
-  });
-  if (!exec) {
-    return callback(null, client, done, logentry);
-  }
-
-  var path = logentry.path;
-  var toData = logentry.data.toData;
-  if (logentry.evtClass === 'Move') {
-    path = logentry.data.toPath;
-    toData = logentry.data.fromData;
-  }
-  var queries = [];
-  if (toData) {
-    if (toData.hasOwnProperty('tags')) {
-      var tagBox = new LinkedDataBox(toData.tags);
-      tagBox.iterateTags(function(pred, tag, idx) {
-        queries.push(
-          generateOneTagInsert(path, pred, tag['@id'], tag.objClass, tag));
-      });
-    }
-  }
-
-  async.each(queries, function(item, callback) {
-    return client.query(item, callback);
-  }, function(err) {
-    return callback(err, client, done, logentry);
-  });
 }
 
 function addActorToLogentry(logentry, access) {
